@@ -1,5 +1,6 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
 using System.Runtime.InteropServices;
+using Robust.Shared;
 using System.Threading;
 using OpenToolkit.GraphicsLibraryFramework;
 using Robust.Client.Input;
@@ -8,6 +9,8 @@ using GlfwKey = OpenToolkit.GraphicsLibraryFramework.Keys;
 using GlfwButton = OpenToolkit.GraphicsLibraryFramework.MouseButton;
 using static Robust.Client.Input.Mouse;
 using static Robust.Client.Input.Keyboard;
+using Robust.Shared.IoC;
+using Robust.Shared.Configuration;
 
 namespace Robust.Client.Graphics.Clyde
 {
@@ -22,6 +25,7 @@ namespace Robust.Client.Graphics.Clyde
 
             private void InitKeyMap()
             {
+                _printableKeyNameMap.Clear();
                 // From GLFW's source code: this is the actual list of "printable" keys
                 // that GetKeyName returns something for.
                 CacheKey(Keys.KeyPadEqual);
@@ -41,8 +45,18 @@ namespace Robust.Client.Graphics.Clyde
                     if (rKey == Key.Unknown)
                         return;
 
-                    var name = GLFW.GetKeyName(key, 0);
-                    if (name != null)
+                    string name;
+
+                    if (!_clyde._cfg.GetCVar(CVars.DisplayUSQWERTYHotkeys))
+                    {
+                        name = GLFW.GetKeyName(key, 0);
+                    }
+                    else
+                    {
+                        name = key.ToString();
+                    }
+
+                    if (!string.IsNullOrEmpty(name))
                         _printableKeyNameMap.Add(rKey, name);
                 }
             }
