@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
@@ -41,12 +42,6 @@ namespace Robust.Server.GameStates
         [Dependency] private readonly IConfigurationManager _configurationManager = default!;
 
         private ISawmill _logger = default!;
-
-        public bool PvsEnabled
-        {
-            get => _configurationManager.GetCVar(CVars.NetPVS);
-            set => _configurationManager.SetCVar(CVars.NetPVS, value);
-        }
 
         public float PvsRange
         {
@@ -122,7 +117,6 @@ namespace Robust.Server.GameStates
             DebugTools.Assert(_networkManager.IsServer);
 
             _pvs.ViewSize = PvsRange * 2;
-            _pvs.CullingEnabled = PvsEnabled;
 
             if (!_networkManager.IsConnected)
             {
@@ -130,6 +124,7 @@ namespace Robust.Server.GameStates
                 _entityManager.CullDeletionHistory(GameTick.MaxValue);
                 _pvs.CullDeletionHistory(GameTick.MaxValue);
                 _mapManager.CullDeletionHistory(GameTick.MaxValue);
+                _pvs.Cleanup(_playerManager.GetAllPlayers());
                 return;
             }
 
@@ -202,6 +197,7 @@ namespace Robust.Server.GameStates
                 }
             });
 
+            _pvs.Cleanup(_playerManager.GetAllPlayers());
             var oldestAck = new GameTick(oldestAckValue);
 
             // keep the deletion history buffers clean
@@ -212,79 +208,6 @@ namespace Robust.Server.GameStates
                 _pvs.CullDeletionHistory(oldestAck);
                 _mapManager.CullDeletionHistory(oldestAck);
             }
-        }
-
-        /// <summary>
-        /// Generates a network entity state for the given entity.
-        /// </summary>
-        /// <param name="entMan">EntityManager that contains the entity.</param>
-        /// <param name="player">The player to generate this state for.</param>
-        /// <param name="entityUid">Uid of the entity to generate the state from.</param>
-        /// <param name="fromTick">Only provide delta changes from this tick.</param>
-        /// <returns>New entity State for the given entity.</returns>
-        internal static EntityState GetEntityState(IServerEntityManager entMan, ICommonSession player, EntityUid entityUid, GameTick fromTick)
-        {
-            var bus = entMan.EventBus;
-            var changed = new List<ComponentChange>();
-
-            foreach (var (netId, component) in entMan.GetNetComponents(entityUid))
-            {
-                DebugTools.Assert(component.Initialized);
-
-                // NOTE: When LastModifiedTick or CreationTick are 0 it means that the relevant data is
-                // "not different from entity creation".
-                // i.e. when the client spawns the entity and loads the entity prototype,
-                // the data it deserializes from the prototype SHOULD be equal
-                // to what the component state / ComponentChange would send.
-                // As such, we can avoid sending this data in this case since the client "already has it".
-
-                DebugTools.Assert(component.LastModifiedTick >= component.CreationTick);
-
-                if (component.CreationTick != GameTick.Zero && component.CreationTick >= fromTick && !component.Deleted)
-                {
-                    ComponentState? state = null;
-                    if (component.NetSyncEnabled && component.LastModifiedTick != GameTick.Zero && component.LastModifiedTick >= fromTick)
-                        state = entMan.GetComponentState(bus, component, player);
-
-                    // Can't be null since it's returned by GetNetComponents
-                    // ReSharper disable once PossibleInvalidOperationException
-                    changed.Add(ComponentChange.Added(netId, state));
-                }
-                else if (component.NetSyncEnabled && component.LastModifiedTick != GameTick.Zero && component.LastModifiedTick >= fromTick)
-                {
-                    changed.Add(ComponentChange.Changed(netId, entMan.GetComponentState(bus, component, player)));
-                }
-            }
-
-            foreach (var netId in entMan.GetDeletedComponents(entityUid, fromTick))
-            {
-                changed.Add(ComponentChange.Removed(netId));
-            }
-
-            return new EntityState(entityUid, changed.ToArray());
-        }
-
-        /// <summary>
-        ///     Gets all entity states that have been modified after and including the provided tick.
-        /// </summary>
-        internal static List<EntityState>? GetAllEntityStates(IServerEntityManager entityMan, ICommonSession player, GameTick fromTick)
-        {
-            var stateEntities = new List<EntityState>();
-            foreach (var entity in entityMan.GetEntities())
-            {
-                if (entity.Deleted)
-                {
-                    continue;
-                }
-
-                DebugTools.Assert(entity.Initialized);
-
-                if (entity.LastModifiedTick >= fromTick)
-                    stateEntities.Add(GetEntityState(entityMan, player, entity.Uid, fromTick));
-            }
-
-            // no point sending an empty collection
-            return stateEntities.Count == 0 ? default : stateEntities;
         }
     }
 }
